// src/app/models/valClientQueueModel.ts
export class ValClientQueueModel {
  id: number;
  date_entry_str: string;
  email: string;
  fio: string;
  first_pay: number;
  date_entry: any;
  iin: string;
  num_dogovor: string;
  phone: string;
  price: number;
  sttus_id: number;
  sum_pay: number;
  term: number;
  type_id: number;
  userId: number;
  ord_status: number;
  proc: number;
  pay: number;

  constructor() {
  }
}